Remember user's stream and volume settings

+++ src/helpers/index.ts
@@ -1,4 +1,4 @@
-import BigNumber from 'bignumber.js'
+import BigNumber from "bignumber.js"
 
 // Capitalize convert a string "example" to "Example"
 export const Capitalize = (str: string): string => str[0].toUpperCase() + str.substring(1).toLowerCase()
@@ -9,7 +9,7 @@ export const Capitalize = (str: string): string => str[0].toUpperCase() + str.su
  * @returns {boolean}
  */
 export const isObject = (item: any) => {
-    return item && typeof item === 'object' && !Array.isArray(item)
+    return item && typeof item === "object" && !Array.isArray(item)
 }
 
 /**
@@ -40,19 +40,19 @@ export const shadeColor = (hexColor: string, percent: number) => {
     let G = parseInt(hexColor.substring(3, 5), 16)
     let B = parseInt(hexColor.substring(5, 7), 16)
 
-    R = parseInt((R * (100 + percent)) / 100 + '')
-    G = parseInt((G * (100 + percent)) / 100 + '')
-    B = parseInt((B * (100 + percent)) / 100 + '')
+    R = parseInt((R * (100 + percent)) / 100 + "")
+    G = parseInt((G * (100 + percent)) / 100 + "")
+    B = parseInt((B * (100 + percent)) / 100 + "")
 
     R = R < 255 ? R : 255
     G = G < 255 ? G : 255
     B = B < 255 ? B : 255
 
-    const RR = R.toString(16).length == 1 ? '0' + R.toString(16) : R.toString(16)
-    const GG = G.toString(16).length == 1 ? '0' + G.toString(16) : G.toString(16)
-    const BB = B.toString(16).length == 1 ? '0' + B.toString(16) : B.toString(16)
+    const RR = R.toString(16).length == 1 ? "0" + R.toString(16) : R.toString(16)
+    const GG = G.toString(16).length == 1 ? "0" + G.toString(16) : G.toString(16)
+    const BB = B.toString(16).length == 1 ? "0" + B.toString(16) : B.toString(16)
 
-    return '#' + RR + GG + BB
+    return "#" + RR + GG + BB
 }
 
 export const getRandomArbitrary = (min: number, max: number): number => {
@@ -61,16 +61,16 @@ export const getRandomArbitrary = (min: number, max: number): number => {
 
 export const supFormatter = (num: BigNumber): string => {
     // sups have 18 decimal places, divide by 1,000,000,000,000,000,000
-    return num.dividedBy(new BigNumber('1000000000000000000')).toFixed()
+    return num.dividedBy(new BigNumber("1000000000000000000")).toFixed()
 }
 
 export const parseString = (val: string | null, defaultVal: number): number => {
     if (!val) return defaultVal
 
-    return parseInt(val)
+    return parseFloat(val)
 }
 
-export const getObjectFromArrayByKey = (array: any[], idValue: string, idName = 'id') => {
+export const getObjectFromArrayByKey = (array: any[], idValue: string, idName = "id") => {
     for (let i = 0; i < array.length; i++) {
         if (array[i][idName] === idValue) {
             return array[i]
